requirements.txt
pandas
numpy
scikit-learn
transformers
joblib
torch
language_tool_python
reportlab
streamlit
PyPDF2
pytesseract
Pillow
python-docx
regex
